enable_camera_mac/pyproject.toml
[tool.poetry]
name = "pycontrol"
version = "0.1.0"
description = "Python control tools for Xreal One Pro"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
hidapi = "^0.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
